Fixes of the project structure
