--- experiments/selenium-java-page-factory/src/test/java/pageObjects/HomePage.java
@@ -1,15 +1,0 @@
-package pageObjects;
-
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.FindBy;
-
-public class HomePage {
-
-    @FindBy (id = "content")
-    static WebElement optionsContent;
-
-    public static void clickOption(String optionText) {
-        optionsContent.findElement(By.linkText(optionText)).click();
-    }
-}
